+++ src/main/java/net/detalk/api/product/domain/ProductMaker.java
@@ -1,4 +1,4 @@
-package net.detalk.api.domain;
+package net.detalk.api.product.domain;
 
 import java.time.Instant;
 import lombok.Builder;
